+++ tests/model/test_modelFunc.py
@@ -412,7 +412,7 @@ class TestModelFuncs (unittest.TestCase):
         self.assertEqual(len(enumDict1), 2)
         self.assertEqual(enumDict1.get("type", None), "string")
         self.assertEqual(enumDict1.get("enum", None), ["topic", "direct", "fanout", "default", "headers"])
-        dictType1 = extractedTypes[26]
+        dictType1 = extractedTypes[25]
         dictDict1 = modelFuncs.typeToJSONDict(dictType1, localTypePrefix)
         self.assertEqual(len(dictDict1), 3)
         self.assertEqual(dictDict1.get("type", None), "object")